refactor(ui): consolidate field mapping section
